otlp-api/requirements.txt
grpcio
opentelemetry-proto
